Move center_crop to BaseImageProcessor (#25122)
